+++ packages/sphinx-test-reports/sphinxcontrib/test_reports/test_reports.py
@@ -62,6 +62,7 @@ def setup(app):
 
     app.add_config_value("tr_suite_id_length", 3, "html")
     app.add_config_value("tr_case_id_length", 5, "html")
+    app.add_config_value("tr_import_encoding", "utf8", "html")
 
     # nodes
     app.add_node(TestResults)